requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
Pillow==10.1.0
python-dateutil==2.8.2
openai==2.8.1
python-dotenv==1.0.0
